% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{filter_links_by_ext}
\alias{filter_links_by_ext}
\title{Filter links by arbitrary extensions}
\usage{
filter_links_by_ext(
  links,
  exts = c("csv", "txt", "xlsx", "sav", "por", "RData")
)
}
\arguments{
\item{links}{Character vector of URLs.}

\item{exts}{Character vector of file extensions (without dot), e.g. c("csv","txt").}
}
\value{
Character vector of filtered URLs.
}
\description{
Filter links by arbitrary extensions
}
