% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{get_all_links}
\alias{get_all_links}
\title{Extract all hyperlinks from a session}
\usage{
get_all_links(session)
}
\arguments{
\item{session}{A polite session object.}
}
\value{
Character vector of absolute URLs.
}
\description{
Extract all hyperlinks from a session
}
