% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{get_file_info}
\alias{get_file_info}
\title{Memoised extractor of file info for various types}
\usage{
get_file_info(url)
}
\arguments{
\item{url}{Character: link to the file.}
}
\value{
A named list: name, url, type, rows, cols or objects.
}
\description{
Memoised extractor of file info for various types
}
