% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{init_session}
\alias{init_session}
\title{Initialize a polite scraping session for a given URL}
\usage{
init_session(url)
}
\arguments{
\item{url}{Character. Root URL to crawl.}
}
\value{
A polite session object.
}
\description{
Initialize a polite scraping session for a given URL
}
